backend/pyproject.toml
[tool.poetry]
name = "boffin"
version = "0.0.0"
description = ""
authors = ["Kris Steinhoff <ksteinhoff@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
sqlalchemy = "^2.0.31"
sqlmodel = "^0.0.19"
fastapi-users = { extras = ["sqlalchemy"], version = "^13.0.0" }
psycopg = { extras = ["binary"], version = "^3.1.19" }
pydantic = "^2.7.4"
strawberry-graphql = { extras = ["fastapi"], version = "^0.235.1" }
alembic = "^1.13.2"
pydantic-settings = "^2.3.4"
redis = "^5.0.7"
structlog = "^24.2.0"
rich = "^13.7.1"
colorama = "^0.4.6"
python-dotenv = "^1.0.1"
result = "^0.17.0"


[tool.poetry.group.dev.dependencies]
typos = "^1.19.0"
ruff = "^0.4"
mypy = "1.7.1"
debugpy = "^1.8.0"
pyparsing = "^3.1.2"
ipykernel = "^6.29"
strawberry-graphql = { extras = ["debug-server"], version = "^0.235.1" }
types-redis = "^4.6.0.20240425"
types-colorama = "^0.4.15.20240311"


[tool.poetry.group.test.dependencies]
pytest = "^8.2.1"
pytest-playwright = "^0.5.0"
pytest-base-url = "^2.0.0"
pytest-cov = "^5.0"
pytest-env = "^1.1.3"
pytest-mock = "^3.12.0"
pytest-xdist = "^3.5.0"
freezegun = "^1.5.1"
semgrep = "^1.64"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pylint.messages_control]
max-line-length = 88

[tool.ruff]
line-length = 88

[tool.pytest.ini_options]
addopts = ["--cov=.", "--cov-report=lcov", "--cov-report=html"]
env = ["DATABASE_URL=sqlite:///db.sqlite3"]

[tool.coverage.report]
exclude_lines = ["if __name__ == .__main__.:", "if TYPE_CHECKING:"]
